Move NmsClassResolver and OcbClassResolver to minecraft package
